Add tracking based on missed days
I added a method to TimeManager that seeks for every day that
does not have entry in the log and handles the list of such days
to the logging view.

I think this is generally a better approach than it was before.
Before you had to keep in mind the last day you logged, which is
not very intuitive. Now you don't have to worry about any days
you missed. The logger takes care of that.

* Changed Press Enter behavior in home_view(). Now it calls
  missed_days_based_logging_view() by default.
* Made small changes to UI [default values in quare brackets].