Temporary fix for *.json files location bug
When you were trying to run main.py from outside the directory - it
wouldn't find *.json files, which are crucial for app. So I quickly fixed
it by adding os.chdir() into views.py. It's not a really good way to fix
that kind of thing, but it's gonna work now.